move component configs to new file
